Machine: correct memory stall cycles computation.
If the access is uncached then then the penalty
to access memory is the access time decreased by one
cycle which is required for the instruction anyway.

When cache is enabled then the first cycle is used
for lookup if data are present in cache and then actual
accesses cost the full penalty of next level memory
access.

Computation is only illustrative. Real memory stalls
can be masked by need for stalls in pipeline (data
hazards or control hazards caused) or the data
stalls can overlap instructions reads stalls
so real effect of the memory stalls is can be
smaller than simple sum of instructions and data
cache computed stalls.

Signed-off-by: Pavel Pisa <pisa@cmp.felk.cvut.cz>